+++ src/main/java/techcourse/myblog/controller/core/MyBlogGlobalMvcConfig.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web.core;
+package techcourse.myblog.controller.core;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -6,7 +6,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
 
-import techcourse.myblog.web.session.SessionArgumentResolver;
+import techcourse.myblog.controller.session.SessionArgumentResolver;
 
 @Configuration
 public class MyBlogGlobalMvcConfig implements WebMvcConfigurer {
