+++ src/main/java/techcourse/myblog/controller/core/UserLoginExceptionAdvice.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web.core;
+package techcourse.myblog.controller.core;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
