+++ src/main/java/techcourse/myblog/repository/ArticleRepository.java
@@ -1,7 +1,9 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import techcourse.myblog.domain.Article;
 
 @Repository
 public interface ArticleRepository extends CrudRepository<Article, Long> {
