+++ src/main/java/techcourse/myblog/repository/UserRepository.java
@@ -1,9 +1,11 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+
+import techcourse.myblog.domain.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
