+++ src/test/java/techcourse/myblog/controller/IndexControllerTest.java
@@ -1,10 +1,10 @@
-package techcourse.myblog.web;
+package techcourse.myblog.controller;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-import techcourse.myblog.web.test.WebClientGenerator;
+import techcourse.myblog.controller.test.WebClientGenerator;
 
 import static org.springframework.http.HttpMethod.GET;
